add forgot password component

+++ app/src/app/components/login/login.component.ts
@@ -2,8 +2,10 @@ import { Component, inject } from '@angular/core';
 import { CommonModule, formatCurrency } from '@angular/common';
 import { RouterModule, Router} from '@angular/router';
 import { AuthServiceService } from '../../services/auth-service.service';
+import { LoginResponse } from '../../models/Books';
 import { FormBuilder, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
+
 
 @Component({
   selector: 'app-login',
@@ -30,14 +32,20 @@ export class LoginComponent {
   onSubmit(){
     if (this.loginForm.valid) {
       this.AuthService.loginUser(this.loginForm.value).subscribe(
-        (response) => {
-           console.log(response.token)
-           localStorage.clear();
-           this.AuthService.login();
-           this.AuthService.setUser(response.user);
-           localStorage.setItem("token", response.user.token);
-           this.router.navigate(['']);
-           this.toastrService.success('Success', 'Login successful');
+        (response : LoginResponse) => {
+           if(response.success)
+           {
+            localStorage.clear();
+            this.AuthService.login();
+            this.AuthService.setUser(response.user);
+            localStorage.setItem("token", response.user.token);
+            this.router.navigate(['']);
+            this.toastrService.success('Success', 'Login successful');
+           }
+           else{
+            this.toastrService.warning(response.message);
+           }
+           
         },
         (error) => {
            console.log(error)
